Pequeñas mejoras

+++ backend_mongoose/app/controllers/concert.controller.js
@@ -1,6 +1,7 @@
 const Concert = require("../models/concert.model.js");
 const Category = require("../models/category.model.js");
 const Artist = require("../models/artist.model.js");
+const Venue = require("../models/artist.model.js");
 
 const getAllConcerts = async (req, res) => {
   try {
@@ -89,13 +90,16 @@ const createConcert = async (req, res) => {
     if (!category) {
       res.status(400).json({ message: "Ha ocurrido un error al buscar la categoria" });
     }
-
-    // const id_artist = req.body.artist;
-    // const artist = await Artist.findOne({ _id: id_artist }).exec();
-    // if (!artist) {
-    //   res.status(400).json({ message: "Ha ocurrido un error al buscar el Artista" });
-    // }
-
+    const id_artist = req.body.artist;
+    const artist = await Artist.findOne({ _id: id_artist }).exec();
+    if (!artist) {
+      res.status(400).json({ message: "Ha ocurrido un error al buscar el Artista" });
+    }
+    const id_venue = req.body.venue;
+    const venue = await Venue.findOne({ _id: id_venue }).exec();
+    if (!venue) {
+      res.status(400).json({ message: "Ha ocurrido un error al buscar la Direccion" });
+    }
     const newConcert = new Concert(concertData);
     await newConcert.save();
     if (!newConcert) {
@@ -115,13 +119,13 @@ const deleteOneConcert = async (req, res) => {
     if (!concert) {
       return res.status(404).json({ message: "Concierto no encontrado" });
     }
-    // const id_cat = concert.category;
-    // const category = await Category.findOne({ _id: id_cat }).exec();
-    // if (!category) {
-    //   return res.status(404).json({ message: "Categoria del concierto no encontrada" });
-    // }
+    const id_cat = concert.category;
+    const category = await Category.findOne({ _id: id_cat }).exec();
+    if (!category) {
+      return res.status(404).json({ message: "Categoria del concierto no encontrada" });
+    }
     await Concert.deleteOne({ _id: concert._id });
-    // await category.removeConcert(concert._id);
+    await category.removeConcert(concert._id);
     return res.status(200).json({ message: "Concierto eliminado" });
   } catch (error) {
     return res.status(500).json({ message: "Error al eliminar el concierto", error: error.message });
